third_party/WebKit/LayoutTests/inspector/console/console-xml-document-expected.txt

CONSOLE MESSAGE: line 9: [object XMLDocument]
Tests that XML document contents are logged using the correct case in the console.

console-xml-document.html:9 #document<MixedCase> Test </MixedCase>
